% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_plot_function.R
\name{indeed_map}
\alias{indeed_map}
\title{Indeed interactive Map}
\usage{
indeed_map(map_ready_df)
}
\arguments{
\item{map_ready_df}{A {Data Frame} result of the output of the function of the same name}
}
\value{

}
\description{
create an interactive map of the indeed data frame created with map_ready_df function.
It displays the job offers' location, its type and its external link to the indeed website
}
\examples{
indeed_data \%>\% map_ready_df() \%>\% indeed_map()
}
