% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Webscrap_Indeed.R
\name{indeed_scraping}
\alias{indeed_scraping}
\title{Indeed web scraping}
\usage{
indeed_scraping(category, number_of_pages)
}
\arguments{
\item{category}{{character vector} with inputed values of the job categories you would like to scrape}
}
\value{

}
\description{
Create a data frame with indeed jobs based on the categories selected.
ATTENTION : Run to update the dataframe containing the information about the jobs listed at Indeed.
Include more or less categories depending on computing power of your computer (Four categories takes about 45 minutes)
}
\examples{
indeed_scraping(category = c("data", "finance", "business"))
}
